move certificate chain parsing to seperate file
